+++ src/plugins/Trader/helpers/index.ts
@@ -1,2 +1,1 @@
 export * from './uniswap'
-export * from './token'
