+++ src/web3/helpers.ts
@@ -1,6 +1,6 @@
 import BigNumber from 'bignumber.js'
 import { getConstant } from './constants'
-import { ChainId } from './types'
+import { ChainId, EthereumTokenType, Token } from './types'
 
 export function isSameAddress(addrA: string, addrB: string) {
     return addrA.toLowerCase() === addrB.toLowerCase()
@@ -24,4 +24,32 @@ export function isUSDT(address: string) {
 
 export function addGasMargin(value: BigNumber) {
     return value.multipliedBy(new BigNumber(10000).plus(new BigNumber(1000))).dividedToIntegerBy(new BigNumber(10000))
+}
+
+export function createEetherToken(chainId: ChainId): Token {
+    return {
+        type: EthereumTokenType.Ether,
+        chainId,
+        address: getConstant('ETH_ADDRESS'),
+        decimals: 18,
+        name: 'Ether',
+        symbol: 'ETH',
+    }
+}
+
+export function createERC20Token(
+    chainId: ChainId,
+    address: string,
+    decimals: number,
+    name: string,
+    symbol: string,
+): Token {
+    return {
+        type: EthereumTokenType.ERC20,
+        chainId,
+        address,
+        decimals,
+        name,
+        symbol,
+    }
 }
